+++ pyqmix/config/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from .directories import DLL_DIR, CONFIG_DIR
-from .pumps import PUMP_1, PUMP_2, PUMP_3, PUMP_4, PUMP_5, PUMP_6
-
-
-__all__ = [DLL_DIR, CONFIG_DIR, PUMP_1, PUMP_2, PUMP_3, PUMP_4, PUMP_5, PUMP_6]