+++ pyqmix/error.py
@@ -10,7 +10,12 @@ from .headers import ERROR_HEADER
 
 class QmixError(object):
     def __init__(self, error_number):
-        self.dll_file = os.path.join(config.DLL_DIR, 'usl.dll')
+        dll_dir = config.read_config().get('qmix_dll_dir', None)
+        if dll_dir is None:
+            self.dll_file = 'usl.dll'
+        else:
+            self.dll_file = os.path.join(dll_dir, 'usl.dll')
+
         self._ffi = FFI()
         self._ffi.cdef(ERROR_HEADER)