+++ pyqmix/headers.py
@@ -27,6 +27,7 @@ PUMP_HEADER = """
     long LCP_LookupPumpByName(const char *pPumpName, dev_hdl* PumpHandle);
     long LCP_IsEnabled(dev_hdl hPump);
     long LCP_Enable(dev_hdl hPump);
+    long LCP_Disable(dev_hdl hPump);
     long LCP_SyringePumpCalibrate(dev_hdl 	hPump);
     long LCP_IsCalibrationFinished(dev_hdl hPump);
     long LCP_IsInFaultState(dev_hdl hPump);  
@@ -38,6 +39,8 @@ PUMP_HEADER = """
                             int    Prefix,
                             int     VolumeUnit,
                             int     TimeUnit);
+    long LCP_GetSyringeParam(dev_hdl hPump, double *pInnerDiameter_mm,
+        double *pMaxPistonStroke_mm);	
     long LCP_SetSyringeParam(dev_hdl hPump,
                                 double  InnerDiameter_mm,
                                 double  MaxPistonStroke_mm);
@@ -53,6 +56,10 @@ PUMP_HEADER = """
     long LCP_GetFlowIs(dev_hdl hPump, double *pFlowRateIs);
     long LCP_IsPumping(dev_hdl hPump);
     long LCP_GetFlowRateMax(dev_hdl hPump, double* FlowRateMax);
+    long LCP_GetFlowUnit(dev_hdl hPump, int *pPrefix, int *pVolumeUnit,
+        int *pTimeUnit);		
+    long LCP_GetVolumeUnit(dev_hdl hPump, int *pPrefix, int *pVolumeUnit);
+    long LCP_GetVolumeMax(dev_hdl hPump, double *VolumeMax)
     long LCP_StopPumping(dev_hdl hPump);
     long LCP_StopAllPumps();
     long LCP_GetValveHandle(dev_hdl hPump, dev_hdl* ValveHandle);